feat(cordic.py): restart the file everytime we get the output.
In the previous versions it was detected that when we give the input it
wait for the output and after that it exit the .py file. we want to give
multiple input after receiving the output for which we update the script
to take the input everytime after the output is received.

Signed-off-by: tej-dabhankar <tejasdabhankar123@gmail.com>